TESTS: move polars column tests into test_column
The pandas and polars case factories drive their tests through different
mechanisms (get_cases_func with option contexts vs. error-expectation
tuples), so the polars tests move verbatim rather than merging into one
driver. Collected test count (81) and coverage are unchanged.
